feat: initial support for rust WIP
